requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
pymongo==4.6.2
groq==0.4.2
httpx==0.27.0
msal==1.27.0